# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Полная документация API для электронного дневника",
        "title": "API системы электронного дневника",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/student-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Получить список групп студентов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID куратора",
                        "name": "curator_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID учебного года",
                        "name": "academic_year_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Ограничение",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service_internal_domain_models.StudentGroup"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Создать группу студентов",
                "parameters": [
                    {
                        "description": "Группа",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.StudentGroup"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.StudentGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/service_internal_lib_api_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service_internal_lib_api_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/student-groups/public": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Получить список публичных групп",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ограничение",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service_internal_domain_models.StudentGroupPublic"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/student-groups/public/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Получить публичную группу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.StudentGroupPublic"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service_internal_lib_api_response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/student-groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Получить группу по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.StudentGroup"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service_internal_lib_api_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Обновить группу студентов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Группа",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.StudentGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service_internal_domain_models.StudentGroup"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service_internal_lib_api_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "student-groups"
                ],
                "summary": "Удалить группу студентов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID группы",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service_internal_lib_api_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service_internal_domain_models.StudentGroup": {
            "type": "object",
            "properties": {
                "academic_year_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "curator_id": {
                    "type": "integer"
                },
                "student_group_id": {
                    "type": "integer"
                },
                "student_group_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "service_internal_domain_models.StudentGroupPublic": {
            "type": "object",
            "properties": {
                "academic_year_id": {
                    "type": "integer"
                },
                "curator_first_name": {
                    "type": "string"
                },
                "curator_id": {
                    "type": "integer"
                },
                "curator_last_name": {
                    "type": "string"
                },
                "curator_middle_name": {
                    "type": "string"
                },
                "student_group_id": {
                    "type": "integer"
                },
                "student_group_name": {
                    "type": "string"
                }
            }
        },
...
        "service_internal_lib_api_response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}